[KED-1659] Delete deprecated kedro.cli and kedro.context modules (#649)
